+++ VehicleTracker.Business/VehicleRevenueLicenceService.cs
@@ -155,8 +155,10 @@ namespace VehicleTracker.Business
         if (!string.IsNullOrEmpty(licenceRecord.Attachment))
         {
           var existingImagePath = string.Format(@"{0}\{1}", folderPath, licenceRecord.Attachment);
-
-          File.Delete(existingImagePath);
+          if (File.Exists(existingImagePath))
+          {
+            File.Delete(existingImagePath);
+          }
         }
 
         if (!Directory.Exists(folderPath))
